refactor(ops): mutualise la redaction PII dans lib/anonymize
createRedactor(anonymize) remplace shortHash/redact* copiés dans 3 scripts.
